Consolidate array_to_string
